server/requirements.txt
django==3.0.6
djangorestframework==3.11.0
mysqlclient==1.4.6
django-extensions==2.2.9

cryptography

tweepy
pyyaml
python-dateutil
nagisa
